modificated comic models
