--- src/app/shared/models/comics/characterList.ts
@@ -1,7 +1,0 @@
-import { CharacterSummary } from "./characterSummary";
-export interface CharacterList{
-    available : number;
-    returned : number;
-    collectionURI : string
-    items : CharacterSummary[];
-}
