--- src/app/shared/models/comics/characterSummary.ts
@@ -1,5 +1,0 @@
-export interface CharacterSummary{
-    resourceURI : string
-    name : string
-    role : string
-}
