--- src/app/shared/models/comics/comicDataContainer.ts
@@ -1,8 +1,0 @@
-import {Comic} from './comics'
-export interface ComicDataContainer{
-    offset : number;
-    limit : number;
-    total : number;
-    count : number;
-    results : Comic[];
-}
