--- src/app/shared/models/comics/comicDataWrapper.ts
@@ -1,10 +1,0 @@
-import {ComicDataContainer} from './ComicDataContainer';
-export interface ComicDataWrapper{
-    code : number;
-    status : string;
-    copyright : string;
-    attributionText : string;
-    attributionHTML : string;
-    data : ComicDataContainer;
-    etag : string;
-}
