--- src/app/shared/models/comics/comicDate.ts
@@ -1,4 +1,0 @@
-export interface ComicDate{
-    type : string;
-    date : Date;
-}
