--- src/app/shared/models/comics/comicPrice.ts
@@ -1,4 +1,0 @@
-export interface ComicPrice{
-    type : string;
-    price : number;
-}
