--- src/app/shared/models/comics/comicSummary.ts
@@ -1,4 +1,0 @@
-export interface ComicSummary{
-    resourceURI : string
-    name : string;
-}
