--- src/app/shared/models/comics/creatorList.ts
@@ -1,7 +1,0 @@
-import { CreatorSummary } from "./creatorSummary";
-export interface CreatorList{
-    available : number;
-    returned : number;
-    collectionURI : string;
-    items : CreatorSummary[];
-}
