--- src/app/shared/models/comics/creatorSummary.ts
@@ -1,5 +1,0 @@
-export interface CreatorSummary{
-    resourceURI : string;
-    name : string
-    role : string
-}
