--- src/app/shared/models/comics/image.ts
@@ -1,4 +1,0 @@
-export interface Image{
-    path : string
-    extension : string
-}
