--- src/app/shared/models/comics/seriesSummary.ts
@@ -1,4 +1,0 @@
-export interface SeriesSummary{
-    resourceURI : string;
-    name : string;
-}
