--- src/app/shared/models/comics/textObject.ts
@@ -1,6 +1,0 @@
-export interface TextObject{
-    type : string;
-    language : string;
-    text : string;
-
-}
