--- src/app/shared/models/comics/url.ts
@@ -1,4 +1,0 @@
-export interface Url{
-    type : string;
-    url : string;
-}
